+++ game_patch/rf/geometry.h
@@ -394,4 +394,6 @@ namespace rf
     static_assert(sizeof(GPortalObject) == 0x30);
 
     static auto& g_cache_clear = addr_as_ref<void()>(0x004F0B90);
+
+    static auto& bbox_intersect = addr_as_ref<bool(const Vector3& bbox1_min, const Vector3& bbox1_max, const Vector3& bbox2_min, const Vector3& bbox2_max)>(0x0046C340);
 }