+++ docs/CHANGELOG.md
@@ -5,6 +5,7 @@ Version 1.8.1 (not released yet)
 --------------------------------
 - Fix possible crash when alt-tabbing in fullscreen mode before game is fully initialized
 - Simplify installation detection in setup and launcher
+- Fix writing corrupted save files when autosaving some levels
 
 Version 1.8.0 (released 2022-09-17)
 -----------------------------------