+++ docs/CHANGELOG.md
@@ -11,6 +11,7 @@ Version 1.8.1 (not released yet)
 - Don't copy over the changelog during setup
 - Add `gamma` command
 - Fix holes in driller cockpit on a widescreen display
+- Change default 'Use' key to E
 
 Version 1.8.0 (released 2022-09-17)
 -----------------------------------