+++ docs/CHANGELOG.md
@@ -72,6 +72,7 @@ Version 1.8.0 (released 2022-09-17)
 - Change default FPS limit to 120 to comply with the stock game
 - Fix loading a save game when player entity is out of level bounds
 - Fix possible freeze when burning entity is destroyed
+- Fix crash reporter URL
 
 Version 1.7.0 (released 2021-06-05)
 -----------------------------------