+++ docs/CHANGELOG.md
@@ -19,6 +19,7 @@ Version 1.8.1 (not released yet)
 - Speed up shadows rendering in levels with multiple detailed movers
 - Add `debug particle_emitter` subcommand (shows emitter cull radius)
 - Fix crash reporter URL
+- Fix buffer-overflow when importing mesh with more than 8000 faces in the editor
 
 Version 1.8.0 (released 2022-09-17)
 -----------------------------------