+++ docs/CHANGELOG.md
@@ -22,6 +22,7 @@ Version 1.8.1 (not released yet)
 - Fix buffer-overflow when importing mesh with more than 8000 faces in the editor
 - Fix various issues when server switches to a new level before player finishes downloading the previous one
 - Adjust letterbox effects in cutscenes and after death for wide screens
+- Add support for bik files (Bink videos) in mods
 
 Version 1.8.0 (released 2022-09-17)
 -----------------------------------