+++ docs/CHANGELOG.md
@@ -25,6 +25,7 @@ Version 1.8.1 (not released yet)
 - Add support for bik files (Bink videos) in mods
 - Hide player riot shield third person model in cutscenes
 - Fix player third person weapon model animations after loading the game from a save file
+- Change default volume to 0.5
 
 Version 1.8.0 (released 2022-09-17)
 -----------------------------------